Updated trigger VacationInsert
Trigger for vacation management. When adding a record to the Vacation table, the Payment field should be calculated automatically. The Payment field must be calculated from the formula S=M/(365-C)*N described in the readme file. To perform this accounting, the VacationInsert trigger was created. With the help of the vacation_Cursor fast forward cursor, you can go through each row of the inserted table and add a value to the Payment field. If null is returned from the EmployeeVacationPerc function, then an error is returned